cleaned up basic presentation tool scene ui
