+++ Assets/HotmaxEditorVR/presentation/common/presentationMain.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:064c5f6fd8bccc65099ba0e217165180b4730538c317903d0d1ca377f3661ec9
-size 737649
+oid sha256:3f411a638ee148852c9823d5e68f65b2ded6d42f8200480b61de5d3911b8b04a
+size 307126
