+++ Assets/HotmaxEditorVR/presentation/common/presentationTools(ZEDHolder).prefab
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c7270d7a40644d8b13e484f9ddedafa555f9e6aefeefad8ee35e2c40f0b5e43f
+size 306743
