+++ Assets/HotmaxEditorVR/presentation/common/presentationTools(ZEDHolder).prefab.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 3146e9c130990f04fb37f2935128ea84
-folderAsset: yes
-timeCreated: 1521154481
+guid: 9e55ccbf50aa01a4c8c3d310b39dbfef
+timeCreated: 1522386071
 licenseType: Pro
-DefaultImporter:
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
