+++ Assets/HotmaxEditorVR/presentation/keyingTool/keyingTool.cs
@@ -11,11 +11,11 @@ public class keyingTool : MonoBehaviour
     private stateManager _stateManagerMutatorRef;
 
     private string keyingDataFile = "keying_data.json";
-    private string dataFilePath;
+    private string keyingDataFilePath;
 
     private GreenScreenManager greenScreenManager;
 
-    public GameObject greenScreenGradient;
+    public GameObject greenScreenColorPicker;
 
     public Renderer currentKeyColor;
 
@@ -25,6 +25,8 @@ public class keyingTool : MonoBehaviour
     public TextMeshPro clipBlackText;
     public TextMeshPro erosionText;
     public TextMeshPro desplillText;
+    public TextMeshPro exposureText;
+    public TextMeshPro gainText;
 
     public GameObject rangeSlider;
     public GameObject smoothingSlider;
@@ -32,64 +34,22 @@ public class keyingTool : MonoBehaviour
     public GameObject clipBlackSlider;
     public GameObject erosionSlider;
     public GameObject despillSlider;
+    public GameObject exposureSlider;
+    public GameObject gainSlider;
 
     private void Start()
     {
-        dataFilePath = Path.Combine(Application.streamingAssetsPath, keyingDataFile);
-
-        stateManager.arModeIsOnEvent += updateARModeIsOn;
         greenScreenManager = FindObjectOfType<GreenScreenManager>();
-
-        if (_arModeIsOn)
+        if (greenScreenManager != null)
         {
-            //set range and smoothness green screen vals to 0
-            greenScreenManager.range = 0;
-            greenScreenManager.smoothness = 0;
-            greenScreenManager.UpdateShader();
+            loadChromaKeyData();
         }
-    }
-
-    private void OnApplicationQuit()
-    {
-        stateManager.arModeIsOnEvent -= updateARModeIsOn;
-    }
-
-    private void OnEnable()
-    {
-        loadChromaKeyData();
-
-        if (_arModeIsOn && greenScreenManager != null)
-        {
-            //set range and smoothness green screen vals to 0
-            greenScreenManager.range = 0;
-            greenScreenManager.smoothness = 0;
-            greenScreenManager.UpdateShader();
-        }
-    }
-
-    //--fires when keying tools toggled off as well as depthcast app shuts down
-    private void OnDisable()
-    {
-        saveChromaKeyData();
     }
 
     void Update()
     {
         setKeyColor();
         setChromaKeyValues();
-    }
-
-
-    void updateARModeIsOn(bool value)
-    {
-        _arModeIsOn = value;
-        if (_arModeIsOn)
-        {
-            //set range and smoothness green screen vals to 0
-            greenScreenManager.range = 0;
-            greenScreenManager.smoothness = 0;
-            greenScreenManager.UpdateShader();
-        }
     }
 
     void setKeyColor()
@@ -101,7 +61,7 @@ public class keyingTool : MonoBehaviour
             if (inputManager.trackedController2.triggerPressed)
             {
                 //--Gets the pixel color from the green screen gradient swatch next to the actor monitor, then applies that to the keyColors of "GreenScreenManager"
-                if (hit.collider.gameObject == greenScreenGradient)
+                if (hit.collider.gameObject == greenScreenColorPicker)
                 {
                     Renderer rend = hit.transform.GetComponent<Renderer>();
                     MeshCollider meshCollider = hit.collider as MeshCollider;
@@ -136,6 +96,10 @@ public class keyingTool : MonoBehaviour
             greenScreenManager.blackClip = clipBlackSlider.transform.localPosition.x;
             greenScreenManager.erosion = Mathf.RoundToInt(erosionSlider.transform.localPosition.x * 5);
             greenScreenManager.spill = despillSlider.transform.localPosition.x;
+
+            sl.ZEDCamera.GetInstance().SetCameraSettings(sl.CAMERA_SETTINGS.EXPOSURE, Mathf.RoundToInt(exposureSlider.transform.localPosition.x * 100));
+            sl.ZEDCamera.GetInstance().SetCameraSettings(sl.CAMERA_SETTINGS.GAIN, Mathf.RoundToInt(gainSlider.transform.localPosition.x * 100));
+
             updateTextValues();
             greenScreenManager.UpdateShader();
         }
@@ -149,21 +113,27 @@ public class keyingTool : MonoBehaviour
         clipBlackSlider.transform.localPosition = new Vector3(greenScreenManager.blackClip, clipBlackSlider.transform.localPosition.y, clipBlackSlider.transform.localPosition.z);
         erosionSlider.transform.localPosition = new Vector3(greenScreenManager.erosion / (float)5, erosionSlider.transform.localPosition.y, erosionSlider.transform.localPosition.z);
         despillSlider.transform.localPosition = new Vector3(greenScreenManager.spill, despillSlider.transform.localPosition.y, despillSlider.transform.localPosition.z);
+
+        exposureSlider.transform.localPosition = new Vector3(sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.EXPOSURE) / (float)100, exposureSlider.transform.localPosition.y, exposureSlider.transform.localPosition.z);
+        gainSlider.transform.localPosition = new Vector3(sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.GAIN) / (float)100, gainSlider.transform.localPosition.y, gainSlider.transform.localPosition.z);
     }
 
     void updateTextValues()
     {
-        rangeText.text = greenScreenManager.range.ToString();
-        smoothingText.text = greenScreenManager.smoothness.ToString();
-        clipWhiteText.text = greenScreenManager.whiteClip.ToString();
-        clipBlackText.text = greenScreenManager.blackClip.ToString();
-        erosionText.text = greenScreenManager.erosion.ToString();
-        desplillText.text = greenScreenManager.spill.ToString();
+        rangeText.text =     "Range      : " + greenScreenManager.range.ToString();
+        smoothingText.text = "Smoothing  : " + greenScreenManager.smoothness.ToString();
+        clipWhiteText.text = "Clip White : " + greenScreenManager.whiteClip.ToString();
+        clipBlackText.text = "Clip Black : " + greenScreenManager.blackClip.ToString();
+        erosionText.text =   "Erosion    : " + greenScreenManager.erosion.ToString();
+        desplillText.text =  "Despill    : " + greenScreenManager.spill.ToString();
+
+        exposureText.text =  "Erosion    : " + sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.EXPOSURE).ToString();
+        gainText.text =      "Gain       : " + sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.GAIN).ToString();
     }
 
     public void saveChromaKeyData()
     {
-        saveChromakeyValues saveData = new saveChromakeyValues();
+        saveChromakeyData saveData = new saveChromakeyData();
         saveData.keyColor = greenScreenManager.keyColors;
         saveData.range = greenScreenManager.range;
         saveData.smoothness = greenScreenManager.smoothness;
@@ -172,20 +142,24 @@ public class keyingTool : MonoBehaviour
         saveData.erosion = greenScreenManager.erosion;
         saveData.despill = greenScreenManager.spill;
 
+        saveData.exposure = sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.EXPOSURE);
+        saveData.gain = sl.ZEDCamera.GetInstance().GetCameraSettings(sl.CAMERA_SETTINGS.GAIN);
+
         //Convert to Json
         string jsonData = JsonUtility.ToJson(saveData);
         //Save Json string
-        File.WriteAllText(dataFilePath, jsonData);
+        File.WriteAllText(keyingDataFilePath, jsonData);
     }
 
     public void loadChromaKeyData()
     {
-        if (File.Exists(dataFilePath))
+        keyingDataFilePath = Path.Combine(Application.streamingAssetsPath, keyingDataFile);
+        if (File.Exists(keyingDataFilePath))
         {
             //Load saved Json
-            string jsonData = File.ReadAllText(dataFilePath);
+            string jsonData = File.ReadAllText(keyingDataFilePath);
             //Convert to Class
-            saveChromakeyValues loadedData = JsonUtility.FromJson<saveChromakeyValues>(jsonData);
+            saveChromakeyData loadedData = JsonUtility.FromJson<saveChromakeyData>(jsonData);
 
             //Display saved data
             greenScreenManager.keyColors = loadedData.keyColor;
@@ -196,6 +170,9 @@ public class keyingTool : MonoBehaviour
             greenScreenManager.erosion = loadedData.erosion;
             greenScreenManager.spill = loadedData.despill;
 
+            sl.ZEDCamera.GetInstance().SetCameraSettings(sl.CAMERA_SETTINGS.EXPOSURE, loadedData.exposure);
+            sl.ZEDCamera.GetInstance().SetCameraSettings(sl.CAMERA_SETTINGS.GAIN, loadedData.gain);
+
             currentKeyColor.material.color = greenScreenManager.keyColors;
 
             updateSliderPositions();
@@ -203,11 +180,36 @@ public class keyingTool : MonoBehaviour
 
             greenScreenManager.UpdateShader();
         }
+        else
+        {
+            writeDefaults();
+            loadChromaKeyData();
+        }
+    }
+
+    void writeDefaults()
+    {
+        saveChromakeyData saveData = new saveChromakeyData();
+        saveData.keyColor = Color.green;
+        saveData.range = 0.30f;
+        saveData.smoothness = 0;
+        saveData.whiteClip = 0;
+        saveData.blackClip = 0;
+        saveData.erosion = 0;
+        saveData.despill = 0;
+
+        saveData.exposure = 50;
+        saveData.gain = 90;
+
+        //Convert to Json
+        string jsonData = JsonUtility.ToJson(saveData);
+        //Save Json string
+        File.WriteAllText(keyingDataFilePath, jsonData);
     }
 }
 
 [SerializeField]
-public class saveChromakeyValues
+public class saveChromakeyData
 {
     public Color keyColor = Color.green;
     public float range = 0;
@@ -216,5 +218,6 @@ public class saveChromakeyValues
     public float blackClip = 0;
     public int erosion = 0;
     public float despill = 0;
+    public int exposure = 0;
+    public int gain = 0;
 }
-
