--- Assets/HotmaxEditorVR/presentation/keyingTool/newKeyingTool.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: d34bf9f30176db941ba6e7e6a8770290
-timeCreated: 1522270961
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
