--- Assets/HotmaxEditorVR/presentation/preprocessing.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: f60cbc5750f159a4a85c90da0f32776f
-folderAsset: yes
-timeCreated: 1521496028
-licenseType: Pro
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
