--- Assets/HotmaxEditorVR/presentation/teleport/screen.mat
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:975903ee98db40de33ad71bf3bf57b020d47144c95604fb8390774c329468cf3
-size 2117
