--- Assets/HotmaxEditorVR/presentation/teleport/screen.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 951b73c526028444faacc4ecf8cd820b
-timeCreated: 1521155582
-licenseType: Pro
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
