--- Assets/HotmaxEditorVR/presentation/teleport/screen.renderTexture.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 5180492632a7cce4e96b4ad13ed4a6d9
-timeCreated: 1521155120
-licenseType: Pro
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
