--- Assets/HotmaxEditorVR/presentation/teleport/teleportPad (1).prefab
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f48bc189b513a80111375798530b698c61eff2b42c0ef0f5963dad6488a9bf75
-size 12902
