--- Assets/HotmaxEditorVR/presentation/teleport/teleportPad (1).prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 3f7cb18f9d92bb24bade3f27c711a90a
-timeCreated: 1521157537
-licenseType: Pro
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
